Split LODs creation from tiles creation & comments as docstrings
Create a new file 'tileset_creation' where all methods about tiles and tileset are moved. Now, lod_tree contains only the methods about LODs
