+++ py3dtilers/Common/__init__.py
@@ -3,13 +3,15 @@ from .object_to_tile import ObjectToTile, ObjectsToTile, ObjectsToTileWithGeomet
 from .tree_with_children_and_parent import TreeWithChildrenAndParent
 from .lod_1 import get_lod1
 from .loa import create_loa
-from .lod_tree import create_tileset
+from .lod_tree import create_lod_tree
+from .tileset_creation import create_tileset
 
 __all__ = ['kd_tree',
-           'ObjectToTile', 
+           'ObjectToTile',
            'ObjectsToTile',
            'ObjectsToTileWithGeometry',
            'TreeWithChildrenAndParent',
            'create_tileset',
            'get_lod1',
-           'create_loa']
+           'create_loa',
+           'create_lod_tree']
